Add internal testing build tag and test
